Add meet-in-the-middle packing and switch to BigDecimal

+++ src/main/java/com/github/andrepnh/packer/algorithm/PackingAlgorithm.java
@@ -2,8 +2,9 @@ package com.github.andrepnh.packer.algorithm;
 
 import com.github.andrepnh.packer.core.Item;
 import com.github.andrepnh.packer.core.Package;
+import java.math.BigDecimal;
 import java.util.Set;
 
 public interface PackingAlgorithm {
-  Package pack(double weightLimit, Set<Item> candidateItems);
+  Package pack(BigDecimal weightLimit, Set<Item> candidateItems);
 }
