+++ src/main/java/com/github/andrepnh/packer/core/Input.java
@@ -2,13 +2,14 @@ package com.github.andrepnh.packer.core;
 
 import com.google.common.base.MoreObjects;
 import com.google.common.collect.ImmutableSet;
+import java.math.BigDecimal;
 
 public class Input {
-  private final int weightLimit;
+  private final BigDecimal weightLimit;
 
   private final ImmutableSet<Item> items;
 
-  public Input(int weightLimit, Iterable<Item> items) {
+  public Input(BigDecimal weightLimit, Iterable<Item> items) {
     this.weightLimit = weightLimit;
     this.items = ImmutableSet.copyOf(items);
   }
@@ -21,7 +22,7 @@ public class Input {
         .toString();
   }
 
-  public int getWeightLimit() {
+  public BigDecimal getWeightLimit() {
     return weightLimit;
   }
